First draft of an automatic s2p script
Functionalized post-processing and visualization code so that it can be easily called into the automatic script which runs through all of the functions while only requiring the user to input their desired suite2p parameters and the file directory. Allows for visualization of multiple concurrent threshold scaling values which we believe have the greatest influence on the pre-processing output.